refactor(state): split AppState god-file into api_stage/stage_state/osc siblings
state/mod.rs was 1117 prod lines, over the 1000-line file-size hard cap (#486).
Extract three cohesive `impl AppState` method groups into new sibling modules,
leaving mod.rs as struct definition + construction + orchestration only:

- api_stage.rs:  presentation cache, group-color resolution, API-stage snapshots
- stage_state.rs: update_stage_state / clear_stage + slide-edit reconciliation
- osc.rs:        OSC settings application + accessors

Pure mechanical move — no logic, signature, or behavior change. Bare-private
methods called from sibling modules widened to pub(super); tests.rs imports
adjusted for the relocated symbols (no assertion changes). mod.rs now 762 prod
lines (under the 800 warning threshold). All 351 presenter-server tests pass
unchanged.

Closes #486

[no-test: pure file-split refactor, behavior unchanged, existing suite is the guard]
